pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Proyecto Tarea 12.07.2024"
authors = ["alestu25 <alestu25@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
scikit-learn = "^1.5.1"
numpy = "^2.0.0"
panda = "^0.3.1"
matplotlib = "^3.9.1"
seaborn = "^0.13.2"
ucimlrepo = "^0.0.7"
pandas = "^2.2.2"
sodapy = "^2.2.0"
python-dotenv = "^1.0.1"
streamlit = "^1.36.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
